+++ context/context.ts
@@ -1,0 +1,6 @@
+import {createContext} from 'react'
+
+export const AppContext = createContext({
+    hash: "",
+    changeHash: (newHash: string) => {}
+})
