+++ pages/resources/index.tsx
@@ -1,24 +1,30 @@
-import type { GetServerSidePropsContext, InferGetServerSidePropsType, NextPage } from "next";
+import type { NextPage } from "next";
 import styles from "../../styles/Home.module.css";
 import Link from "next/link";
-import { Resource } from "../../models/types";
+import { useContext, useEffect, useState } from "react";
+import { AppContext } from "../../context/context";
 
-type ResourcePropsType = InferGetServerSidePropsType<typeof getServerSideProps>;
+const Resources: NextPage = () => {
+  const quizState = useContext(AppContext)
+  const [resources, setResources] = useState([])
 
-const Resources: NextPage<ResourcePropsType> = ({ resources }: ResourcePropsType) => {
+  useEffect(() => {
+    const fetchResources = async () => {
+      if (quizState.hash) {
+        const data  = await (await fetch('/api/resources', {method: 'POST', body: JSON.stringify({data: quizState.hash}),
+        headers: { 'Content-Type': 'application/json'}})).json()
+      setResources(data)
+      }
+    }
+    fetchResources()
+  }, [])
   return (
     <div className={styles.container}>
-      <h1>List of Resources Page</h1>
-      {resources.map((resource) => <div key={resource.id}><Link href={`/resources/${resource.id}`}>{resource.name}</Link> <br/></div>)}
-      <Link href="/">Back to Home</Link>
+      <h1>Results</h1>
+      <div>{resources.map(resource => <div key={resource['id']}><Link href={`resources/${resource['id']}`} >{resource['name']}</Link><br /></div>)}</div>
+      <Link href='/quiz'>Back to Quiz</Link>
     </div>
-  );
-};
-
-export const getServerSideProps = async (context: GetServerSidePropsContext) => {
-  const result = await fetch(`http://localhost:3000/api/resources?ids=${context.query?.ids}`)
-  const resources : Resource[] = await result.json();
-  return { props: { resources } };
+  )
 };
 
 export default Resources;
